# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: University
  description: University specification API
  version: 1.0.0
  contact:
    name: API Botalov Anton
    email: botalovanton@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
...
  /student:
    post:
      tags:
        - student
      summary: Add student
      description: Allows you to add a new student to the database
      operationId: addStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/studentDto'
        required: true
      responses:
        201:
          description: student created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'

    get:
      tags:
        - student
      summary: All students
      description: Returns a list of all students
      operationId: getAllStudents
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/studentDto'
  /student/{id}:
    get:
      tags:
        - student
      summary: find student
      description: get student by id
      operationId: getStudentById
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/studentDto'
        404:
          description: student not found
          content:
            text/plain:
              schema:
                type: string
                example: 'student with id 999 not found'

    delete:
      tags:
        - student
      summary: delete student
      description: delete an existing student
      operationId: deleteStudent
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        204:
          description: successful student deletion

    put:
      tags:
        - student
      summary: update student
      description: update an existing student
      operationId: updateStudent
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/studentDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /student/filter:
    get:
      tags:
        - student
      summary: find by filter
      description: get a list of students by specified filters
      operationId: getFilteredStudents
      parameters:
        - name: studentId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: studentFirstName
          required: false
          schema:
            type: string
            example: 'Ivan'
          in: query
        - name: studentLastName
          required: false
          schema:
            type: string
            example: 'Ivanov'
          in: query
        - name: groupId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: groupName
          required: false
          schema:
            type: string
            example: 'QWE-365'
          in: query
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/studentDto'
  /teacher:
    post:
      tags:
        - teacher
      summary: Add teacher
      description: Allows you to add a new teacher to the database
      operationId: addTeacher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/teacherDto'
        required: true
      responses:
        201:
          description: teacher created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'

    get:
      tags:
        - teacher
      summary: All teachers
      description: Returns a list of all teachers
      operationId: getAllTeachers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/teacherDto'
  /teacher/{id}:
    get:
      tags:
        - teacher
      summary: find teacher
      description: get teacher by id
      operationId: getTeacherById
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/teacherDto'
        404:
          description: teacher not found
          content:
            text/plain:
              schema:
                type: string
                example: 'teacher with id 999 not found'

    delete:
      tags:
        - teacher
      summary: delete teacher
      description: delete an existing teacher
      operationId: deleteTeacher
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        204:
          description: successful teacher deletion

    put:
      tags:
        - teacher
      summary: update teacher
      description: update an existing teacher
      operationId: updateTeacher
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/teacherDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /teacher/filter:
    get:
      tags:
        - teacher
      summary: find by filter
      description: get a list of teachers by specified filters
      operationId: getFilteredTeachers
      parameters:
        - name: teacherId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: teacherFirstName
          required: false
          schema:
            type: string
            example: 'Ivan'
          in: query
        - name: teacherLastName
          required: false
          schema:
            type: string
            example: 'Ivanov'
          in: query
        - name: departmentId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: departmentName
          required: false
          schema:
            type: string
            example: 'Department of economics'
          in: query
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/teacherDto'
  /lesson:
    post:
      tags:
        - lesson
      summary: Add lesson
      description: Allows you to add a new lesson to the database
      operationId: addLesson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/lessonDto'
        required: true
      responses:
        201:
          description: lesson created
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'

    get:
      tags:
        - lesson
      summary: All lessons
      description: Returns a list of all lessons
      operationId: getAllLessons
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/lessonDto'
  /lesson/{id}:
    get:
      tags:
        - lesson
      summary: find lesson
      description: get lesson by id
      operationId: getLessonById
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/lessonDto'
        404:
          description: lesson not found
          content:
            text/plain:
              schema:
                type: string
                example: 'lesson with id 999 not found'

    delete:
      tags:
        - lesson
      summary: delete lesson
      description: delete an existing lesson
      operationId: deleteLesson
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      responses:
        204:
          description: successful lesson deletion

    put:
      tags:
        - lesson
      summary: update lesson
      description: update an existing lesson
      operationId: updateLesson
      parameters:
        - name: id
          required: true
          schema:
            type: string
            example: '999'
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/lessonDto'
        required: true
      responses:
        204:
          description: successful operation
        400:
          description: validation exception
          content:
            text/plain:
              schema:
                type: string
                example: 'error message'
  /lesson/filter:
    get:
      tags:
        - lesson
      summary: find by filter
      description: get a list of lesson by specified filters
      operationId: getFilteredLessons
      parameters:
        - name: lessonId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: minStartTime
          required: false
          schema:
            type: string
            pattern: yyyy-MM-ddTHH:mm:ssZ
            example: '2999-01-01T09:00:00Z'
          in: query
        - name: maxStartTime
          required: false
          schema:
            type: string
            pattern: yyyy-MM-ddTHH:mm:ssZ
            example: '2999-01-01T10:30:00Z'
          in: query
        - name: courseId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: groupId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: teacherId
          required: false
          schema:
            type: string
            example: '999'
          in: query
        - name: classroom
          required: false
          schema:
            type: string
            example: '123'
          in: query
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/lessonDto'
components:
  schemas:
...
    studentDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          uniqueItems: true
          example: 999
        firstName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivanov'
        lastName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivanov'
        groupId:
          type: integer
          minItems: 1
          example: 999
        groupName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'QWE-345'
    teacherDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          example: 999
        firstName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivan'
        lastName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivanov'
        departmentId:
          type: integer
          minItems: 1
          example: 999
        departmentName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Department of economics'
    lessonDto:
      type: object
      properties:
        id:
          type: integer
          minItems: 1
          uniqueItems: true
          example: 999
        start:
          type: string
          pattern: yyyy-MM-ddTHH:mm:ssZ
          description: greater than current date and time
          example: '2999-01-01T09:00:00Z'
        end:
          type: string
          pattern: yyyy-MM-ddTHH:mm:ssZ
          example: '2999-01-01T10:30:00Z'
        courseId:
          type: integer
          minItems: 1
          example: 999
        courseName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Programming'
        groupId:
          type: integer
          minItems: 1
          example: 999
        groupName:
          type: string
          minLength: 3
          maxLength: 10
          example: 'QWE-345'
        teacherId:
          type: integer
          minItems: 1
          example: 999
        teacherFirstName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivan'
        teacherLastName:
          type: string
          minLength: 3
          maxLength: 15
          example: 'Ivanov'
        classroom:
          type: integer
          minItems: 1
          example: 123